Add files via upload
# Windows Update Simulation

This project is an HTML file that simulates a Windows update process. When opened in a web browser, it displays a loading animation and a message indicating that the system is working on updates. The progress of the update process is shown, along with instructions not to turn off the PC during the update. Users can customize the background color and are informed that the PC will restart several times during the update simulation.